Faction updating Adding values to a reputation Added reputation command handler

+++ Source/NexusForever.WorldServer/Database/Character/Model/CharacterReputation.cs
@@ -7,7 +7,7 @@ namespace NexusForever.WorldServer.Database.Character.Model
     {
         public ulong Id { get; set; }
         public uint FactionId { get; set; }
-        public ulong Value { get; set; }
+        public uint Value { get; set; }
 
         public Character Character { get; set; }
     }
